pyproject.toml
[project]
name = "geomltoolkits"
version = "1.0.0"
description = "Toolkits for Geospatial Machine Learning"
authors = [
    {name = "kshitijrajsharma", email = "skshitizraj@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11",
    "geopandas>=0.14",
    "mercantile>=1.2",
    "numpy>=1.26",
    "pillow>=11.0",
    "rasterio>=1.4",
    "rich>=14.0",
    "rtree>=1.3",
    "pyyaml>=6.0",
    "scipy>=1.14",
    "shapely>=1.8",
    "typer>=0.15",
]

[dependency-groups]
dev = [
    "commitizen>=3.31.0",
    "pre-commit>=4.0",
    "pytest>=8.3.5",
    "ruff>=0.15",
]

[tool.hatch.build.targets.wheel]
packages = ["src/geomltoolkits"]

[project.scripts]
geoml = "geomltoolkits.cli:app"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF"]

[tool.ty.environment]
python-version = "3.10"

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
tag_format = "$version"
version_scheme = "semver"
update_changelog_on_bump = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
